microservices/login_service/requirements.txt
flask
flask-cors
PyJWT
cryptography